backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
geoalchemy2==0.15.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation and settings
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# Redis for caching/sessions
redis==5.2.1

# AWS S3 / Cloudinary for media
boto3==1.35.86
cloudinary==1.42.0

# Stripe for payments
stripe==11.4.1

# HTTP client
httpx==0.28.1

# Background tasks
celery==5.4.0

# PDF generation
weasyprint==63.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Utilities
python-slugify==8.0.4
phonenumbers==8.13.52
